modify: roles APIs
