+++ routes/api.php
@@ -1,8 +1,7 @@
 <?php
 
 use App\Modules\Auth\Http\Controllers\Api\AuthController;
-use App\Modules\RolesAndPermissions\Http\Controllers\Api\RolesAndPermissionsController;
-
+use App\Modules\RolesAndPermissions\Http\Controllers\Api\RoleController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -21,7 +20,6 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-
 Route::prefix('/v1/auth')->name('api.auth.')->group(function () {
     Route::post('/login', [AuthController::class, 'login'])->name('login');
 
@@ -30,11 +28,6 @@ Route::prefix('/v1/auth')->name('api.auth.')->group(function () {
     });
 });
 
-
-Route::prefix('/v1/roles_permissions')->group(function () {
-    Route::get('/', [RolesAndPermissionsController::class, 'index']);
-    Route::post('/', [RolesAndPermissionsController::class, 'store']); 
-    Route::get('/{role}', [RolesAndPermissionsController::class, 'show']); 
-    Route::put('/{role}', [RolesAndPermissionsController::class, 'update']); 
-    Route::delete('/{role}', [RolesAndPermissionsController::class, 'destroy']); 
+Route::prefix('/v1')->group(function () {
+    Route::resource('roles', RoleController::class);
 });
